+++ node/resolvers/logistics/index.ts
@@ -1,11 +1,19 @@
-import { headers, withAuthToken } from '../headers'
-import httpResolver from '../httpResolver'
-import paths from '../paths'
+import { resolvers as pickupResolvers } from './PickupPoint'
+
+import { NearPickupPointsArgs, PickupPointArgs } from './types'
+
+export const fieldResolvers = {
+  ...pickupResolvers,
+}
 
 export const queries = {
-  logistics: httpResolver({
-    headers: withAuthToken(headers.json),
-    method: 'GET',
-    url: account => paths.logisticsConfig(account).shipping,
-  }),
+  logistics: (root, _, {dataSources: {logistics}}: Context) => logistics.shipping(),
+
+  nearPickupPoints: (
+    root, 
+    { lat, long, maxDistance }: NearPickupPointsArgs,
+    {dataSources: {logistics}}: Context
+    ) => logistics.nearPickupPoints(lat, long, maxDistance),
+
+  pickupPoint: (root, { id }: PickupPointArgs, {dataSources: {logistics}}: Context) => logistics.pickupById(id),
 }