+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.52.0] - 2019-02-28
+### Added
+- added pickupPoints query
+- added updateOrderFormCheckin mutation
+
 ## [2.51.0] - 2019-02-27
 ### Added
 - Create assembly options resolvers to `OrderFormItem`